font-patcher: Invent name base "Nerd Font Propo"
[why]
We have three ways to create a patched font
- 'normal' (i.e. monospaced oversized symbols): Nerd Font
- 'mono' (i.e. strictly monospaced small symbols): Nerd Font Mono
- 'proportional' (i.e. proportional symbols): Nerd Font Propo

The first two variants have a name, the third one has no name, it is
also called 'Nerd Font' (and it was THE font, for v2.1.0 (only)).

[how]
Invent and set new name, so that all variants have different names.
Makes refering to them explicit.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>